% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravityshop-package.R
\docType{package}
\name{gravityshop-package}
\alias{gravityshop}
\alias{gravityshop-package}
\title{gravityshop: Resources for Teaching a Workshop on Spatial Interaction Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Templates and data for teaching a workshop on spatial interaction analysis using `R`.
}
\references{
\url{https://github.com/paezha/gravityshop}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paezha/gravityshop}
  \item \url{https://paezha.github.io/gravityshop/}
  \item Report bugs at \url{https://github.com/paezha/gravityshop/issues}
}

}
\author{
\strong{Maintainer}: Antonio Paez \email{paezha@mcmaster.ca} (\href{https://orcid.org/0000-0001-6912-9919}{ORCID})


Antonio Paez, School of Earth, Environment and Society, McMaster University \email{paezha@mcmaster.ca}
}
\keyword{internal}
