% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravityshop-package.R
\name{gs_session_1}
\alias{gs_session_1}
\title{Spatial Interaction Modelling Workshop Session 1 Template (Italian Edition).}
\format{
An R Markdown file
}
\usage{
gs_session_1()
}
\description{
Template for creating a computational notebook for Session 1 of the Italian
edition of the workshop.
}
\keyword{datasets}
