% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravityshop-package.R
\name{gs_session_2}
\alias{gs_session_2}
\title{Spatial Interaction Analysis Workshop Session 2 Template.}
\format{
An R Markdown file
}
\usage{
gs_session_2()
}
\description{
Template for creating a computational notebook for Session 2 of the workshop.
}
\keyword{datasets}
