% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravityshop-package.R
\name{gs_session_3}
\alias{gs_session_3}
\title{Spatial Interaction Analysis Workshop Session 3 Template.}
\format{
An R Markdown file
}
\usage{
gs_session_3()
}
\description{
Template for creating a computational notebook for Session 3 of the workshop.
}
\keyword{datasets}
