% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravityshop-package.R
\name{gs_session_4}
\alias{gs_session_4}
\title{Spatial Interaction Analysis Workshop Session 4 Template.}
\format{
An R Markdown file
}
\usage{
gs_session_4()
}
\description{
Template for creating a computational notebook for Session 4 of the workshop.
}
\keyword{datasets}
