% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravityshop-package.R
\name{gs_session_6}
\alias{gs_session_6}
\title{Spatial Interaction Analysis Workshop Session 6 Template.}
\format{
An R Markdown file
}
\usage{
gs_session_6()
}
\description{
Template for creating a computational notebook for Session 6 of the workshop.
}
\keyword{datasets}
