% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravityshop-package.R
\name{gs_session_7}
\alias{gs_session_7}
\title{Spatial Interaction Analysis Workshop Session 7 Template.}
\format{
An R Markdown file
}
\usage{
gs_session_7()
}
\description{
Template for creating a computational notebook for Session 7 of the workshop.
}
\keyword{datasets}
